pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = "Logistics Telegram Bot"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "aiogram>=3.4.0",
    "telethon>=1.36.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.1.0",
    "apscheduler>=3.10.0",
    "reportlab>=4.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.0",
    "python-jose[cryptography]>=3.3.0",
    "groq>=0.4.0",
    "psutil>=5.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
